using grayscale

+++ ask_3/stage_1.2_bitmap/mandel-lib.c
@@ -27,7 +27,7 @@
  * and uses the escape time algorithm to return a color value
  * used to draw the Mandelbrot Set.
  */
-int mandel_iterations_at_point(double x, double y, int max)
+double mandel_iterations_at_point(double x, double y, int max)
 {
 	double x0 = x;
 	double y0 = y;
@@ -43,6 +43,7 @@ int mandel_iterations_at_point(double x, double y, int max)
 		++iter;
 	}
 
+	//return iter+1-log(log(sqrt(x*x+y*y)))/log(2);
 	return iter;
 }
